Adicionado conversor redrawing-java
Criado classe Data2Java capaz de converter classes do tipo Data para classes Java
